# src/Backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Aluguei API
  description: Sistema de gerenciamento de aluguéis - API completa para proprietários, inquilinos, propriedades, contratos e pagamentos
  version: 1.0.0
  contact:
    name: Equipe Aluguei
    email: suporte@aluguei.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desenvolvimento
  - url: https://api.aluguei.com/v1
    description: Servidor de produção
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login do usuário
      description: Autentica um usuário (proprietário) e retorna token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================
  # PROPRIETÁRIOS (OWNERS)
  # =============================================
  /owners:
    post:
      tags:
        - Owners
      summary: Criar novo proprietário
      description: Cria um novo registro de proprietário no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCreate'
      responses:
        '201':
          description: Proprietário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Owners
      summary: Listar todos os proprietários
      description: Retorna uma lista de todos os proprietários cadastrados
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista de proprietários retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Owner'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /owners/{id}:
    get:
      tags:
        - Owners
      summary: Buscar proprietário por ID
      description: Retorna os detalhes de um proprietário específico
      parameters:
        - $ref: '#/components/parameters/OwnerId'
      responses:
        '200':
          description: Proprietário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Owners
      summary: Atualizar proprietário
      description: Atualiza os dados de um proprietário existente
      parameters:
        - $ref: '#/components/parameters/OwnerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerUpdate'
      responses:
        '200':
          description: Proprietário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Owners
      summary: Deletar proprietário
      description: Remove um proprietário do sistema
      parameters:
        - $ref: '#/components/parameters/OwnerId'
      responses:
        '204':
          description: Proprietário deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /owners/email/{email}:
    get:
      tags:
        - Owners
      summary: Buscar proprietário por email
      description: Retorna um proprietário baseado no email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
          description: Email do proprietário
      responses:
        '200':
          description: Proprietário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================
  # PROPRIEDADES (PROPERTIES)
  # =============================================
  /properties:
    post:
      tags:
        - Properties
      summary: Criar nova propriedade
      description: Cria um novo registro de propriedade para um proprietário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreate'
      responses:
        '201':
          description: Propriedade criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          description: Proprietário não encontrado
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Properties
      summary: Listar todas as propriedades
      description: Retorna uma lista de todas as propriedades cadastradas
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PropertyStatus'
          description: Filtrar por status da propriedade
      responses:
        '200':
          description: Lista de propriedades retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /properties/{id}:
    get:
      tags:
        - Properties
      summary: Buscar propriedade por ID
      description: Retorna os detalhes de uma propriedade específica
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Propriedade encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Properties
      summary: Atualizar propriedade
      description: Atualiza os dados de uma propriedade existente
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdate'
      responses:
        '200':
          description: Propriedade atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Properties
      summary: Deletar propriedade
      description: Remove uma propriedade do sistema (apenas se não houver contratos ativos)
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '204':
          description: Propriedade deletada com sucesso
        '400':
          description: Não é possível deletar propriedade com contratos ativos
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /properties/owner/{ownerId}:
    get:
      tags:
        - Properties
      summary: Listar propriedades por proprietário
      description: Retorna todas as propriedades de um proprietário específico
      parameters:
        - $ref: '#/components/parameters/OwnerIdPath'
      responses:
        '200':
          description: Lista de propriedades do proprietário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================
  # INQUILINOS (TENANTS)
  # =============================================
  /tenants:
    post:
      tags:
        - Tenants
      summary: Criar novo inquilino
      description: Cria um novo registro de inquilino para um proprietário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Inquilino criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Tenants
      summary: Listar todos os inquilinos
      description: Retorna uma lista de todos os inquilinos cadastrados
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista de inquilinos retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tenants/{id}:
    get:
      tags:
        - Tenants
      summary: Buscar inquilino por ID
      description: Retorna os detalhes de um inquilino específico
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Inquilino encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tenants
      summary: Atualizar inquilino
      description: Atualiza os dados de um inquilino existente
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: Inquilino atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tenants
      summary: Deletar inquilino
      description: Remove um inquilino do sistema (apenas se não houver contratos ativos)
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '204':
          description: Inquilino deletado com sucesso
        '400':
          description: Não é possível deletar inquilino com contratos ativos
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tenants/owner/{ownerId}:
    get:
      tags:
        - Tenants
      summary: Listar inquilinos por proprietário
      description: Retorna todos os inquilinos de um proprietário específico
      parameters:
        - $ref: '#/components/parameters/OwnerIdPath'
      responses:
        '200':
          description: Lista de inquilinos do proprietário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================
  # CONTRATOS (CONTRACTS)
  # =============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Número da página
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Número de itens por página
    OwnerId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID do proprietário
    OwnerIdPath:
      name: ownerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID do proprietário
    PropertyId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID da propriedade
...
    TenantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID do inquilino
...
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictError:
      description: 'Conflito (ex: registro duplicado)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          enum: [NOT_FOUND, INVALID_INPUT, ALREADY_EXISTS, UNAUTHORIZED, FORBIDDEN, INTERNAL_ERROR, VALIDATION_ERROR, BUSINESS_RULE, DATABASE_ERROR]
          description: Código do erro
        message:
          type: string
          description: Mensagem de erro descritiva
        details:
          type: object
          description: Detalhes adicionais do erro
      required:
        - code
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Página atual
        limit:
          type: integer
          description: Itens por página
        total:
          type: integer
          description: Total de itens
        pages:
          type: integer
          description: Total de páginas

    # =============================================
    # AUTENTICAÇÃO
    # =============================================
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "proprietario@exemplo.com"
        password:
          type: string
          format: password
          example: "senha123"
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT para autenticação
        expires_at:
          type: string
          format: date-time
          description: Data de expiração do token
        owner:
          $ref: '#/components/schemas/Owner'

    # =============================================
    # PROPRIETÁRIOS
    # =============================================
    Owner:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao.silva@exemplo.com"
        phone:
          type: string
          example: "+5511999999999"
        cpf:
          type: string
          example: "123.456.789-00"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1980-01-15"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    OwnerCreate:
      type: object
      properties:
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao.silva@exemplo.com"
        phone:
          type: string
          example: "+5511999999999"
        password:
          type: string
          format: password
          example: "senha123"
        cpf:
          type: string
          example: "12345678900"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1980-01-15"
      required:
        - name
        - email
        - phone
        - password
        - cpf
    OwnerUpdate:
      type: object
      properties:
        name:
          type: string
          example: "João Silva Santos"
        email:
          type: string
          format: email
          example: "joao.silva@exemplo.com"
        phone:
          type: string
          example: "+5511999998888"
        cpf:
          type: string
          example: "12345678900"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1980-01-15"

    # =============================================
    # PROPRIEDADES
    # =============================================
    PropertyStatus:
      type: string
      enum: [available, rented, maintenance, inactive]
      example: "available"
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Apartamento Centro - 2 Quartos"
        description:
          type: string
          example: "Apartamento bem localizado no centro da cidade"
        address:
          type: string
          example: "Rua das Flores, 123 - Centro"
        city:
          type: string
          example: "São Paulo"
        state:
          type: string
          example: "SP"
        zip_code:
          type: string
          example: "01234-567"
        rent_amount:
          type: number
          format: float
          example: 1500.00
        area:
          type: integer
          example: 70
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/PropertyStatus'
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/Owner'
        contracts:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
    PropertyCreate:
      type: object
      properties:
        title:
          type: string
          example: "Apartamento Centro - 2 Quartos"
        description:
          type: string
          example: "Apartamento bem localizado no centro da cidade"
        address:
          type: string
          example: "Rua das Flores, 123 - Centro"
        city:
          type: string
          example: "São Paulo"
        state:
          type: string
          example: "SP"
        zip_code:
          type: string
          example: "01234567"
        rent_amount:
          type: number
          format: float
          example: 1500.00
        area:
          type: integer
          example: 70
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/PropertyStatus'
        owner_id:
          type: string
          format: uuid
      required:
        - title
        - address
        - city
        - state
        - rent_amount
        - owner_id
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
          example: "Apartamento Centro - 2 Quartos Reformado"
        description:
          type: string
          example: "Apartamento bem localizado no centro da cidade, recentemente reformado"
        address:
          type: string
          example: "Rua das Flores, 123 - Centro"
        city:
          type: string
          example: "São Paulo"
        state:
          type: string
          example: "SP"
        zip_code:
          type: string
          example: "01234567"
        rent_amount:
          type: number
          format: float
          example: 1600.00
        area:
          type: integer
          example: 70
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/PropertyStatus'

    # =============================================
    # INQUILINOS
    # =============================================
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Maria Santos"
        email:
          type: string
          format: email
          example: "maria.santos@exemplo.com"
        phone:
          type: string
          example: "+5511988887777"
        cpf:
          type: string
          example: "98765432100"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1990-05-20"
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/Owner'
        contracts:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
    TenantCreate:
      type: object
      properties:
        name:
          type: string
          example: "Maria Santos"
        email:
          type: string
          format: email
          example: "maria.santos@exemplo.com"
        phone:
          type: string
          example: "+5511988887777"
        cpf:
          type: string
          example: "98765432100"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1990-05-20"
        owner_id:
          type: string
          format: uuid
      required:
        - name
        - email
        - phone
        - cpf
        - owner_id
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Maria Santos Oliveira"
        email:
          type: string
          format: email
          example: "maria.oliveira@exemplo.com"
        phone:
          type: string
          example: "+5511988886666"
        cpf:
          type: string
          example: "98765432100"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1990-05-20"

    # =============================================
    # CONTRATOS
    # =============================================
    ContractStatus:
      type: string
      enum: [active, expired, cancelled, pending]
      example: "active"
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        start_date:
          type: string
          format: date
          example: "2024-01-01"
        end_date:
          type: string
          format: date
          example: "2024-12-31"
        monthly_rent:
          type: number
          format: float
          example: 1500.00
        payment_due_day:
          type: integer
          minimum: 1
          maximum: 28
          example: 5
        status:
          $ref: '#/components/schemas/ContractStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        property:
          $ref: '#/components/schemas/Property'
        tenant:
          $ref: '#/components/schemas/Tenant'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
...
    PaymentStatus:
      type: string
      enum: [pending, paid, overdue, cancelled]
      example: "pending"
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        amount:
          type: number
          format: float
          example: 1500.00
        due_date:
          type: string
          format: date
          example: "2024-01-05"
        paid_amount:
          type: number
          format: float
          nullable: true
          example: 1500.00
        paid_date:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        contract:
          $ref: '#/components/schemas/Contract'
...